+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/AutowiredContainer74.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Symfony\Component\DependencyInjection\Tests\Fixtures;
+
+use Psr\Container\ContainerInterface;
+
+final class AutowiredContainer74
+{
+    /**
+     * @required
+     */
+    public ContainerInterface $container;
+}